modificado:    ErtecMOSTOT/Ertec/WebContent/paginas/informeservice.xhtml 	modificado:    ErtecMOSTOT/Ertec/WebContent/resources/css/estilo.css 	modificado:    ErtecMOSTOT/Ertec/src/mb_Proveedores.java 	modificado:    ErtecMOSTOT/Ertec/src/mbean/mb_Reclamo.java 	modificado:    ErtecMOSTOT/Ertec/src/model/DAO/DAO_Reclamo.java 	modificado:    ErtecMOSTOT/Ertec/src/model/DAO/DAO_infoService.java

+++ ErtecMOSTOT/Ertec/src/model/DAO/DAO_infoService.java
@@ -6,7 +6,7 @@ public class DAO_infoService {
     private String equipo="";
     private long visitas;
     private long tiempoRespuesta;
-    private long tiempoHoras;
+    private double tiempoHoras;
     private long horasAntel;
     private long horasConmutador;
     private long horasRed;
@@ -40,10 +40,12 @@ public class DAO_infoService {
     public void setTiempoRespuesta(long tiempoRespuesta) {
       this.tiempoRespuesta = tiempoRespuesta;
     }
-    public long getTiempoHoras() {
+
+
+    public double getTiempoHoras() {
       return tiempoHoras;
     }
-    public void setTiempoHoras(long tiempoHoras) {
+    public void setTiempoHoras(double tiempoHoras) {
       this.tiempoHoras = tiempoHoras;
     }
     public long getHorasAntel() {
